refactor: move 'constants' path
