+++ Composer/packages/client/src/components/TriggerCreationModal/TriggerCreationModal.tsx
@@ -25,11 +25,11 @@ import {
   getEventTypes,
   getActivityTypes,
 } from '../../utils/dialogUtil';
-import { customEventKey } from '../../utils/constants';
 import { projectIdState, schemasState } from '../../recoilModel/atoms/botState';
 import { userSettingsState } from '../../recoilModel';
 import { validatedDialogsSelector } from '../../recoilModel/selectors/validatedDialogs';
 
+import { customEventKey } from './constants';
 import { getTriggerTypes } from './getTriggerTypes';
 import { styles, dialogWindow, dropdownStyles, intent } from './style';
 import {
