+++ Composer/packages/client/src/components/TriggerCreationModal/validators.ts
@@ -5,8 +5,9 @@ import formatMessage from 'format-message';
 import { luIndexer, combineMessage } from '@bfc/indexers';
 
 import { TriggerFormData, TriggerFormDataErrors } from '../../utils/dialogUtil';
-import { eventTypeKey, customEventKey, intentTypeKey, activityTypeKey } from '../../utils/constants';
 import { nameRegex } from '../../constants';
+
+import { eventTypeKey, customEventKey, intentTypeKey, activityTypeKey } from './constants';
 
 export const getLuDiagnostics = (intent: string, triggerPhrases: string) => {
   const content = `#${intent}\n${triggerPhrases}`;
